requirements.txt
# ===============================
# Web Framework
# ===============================
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0

# ===============================
# Validation & Settings
# ===============================
pydantic==2.5.0
pydantic-core==2.14.1
pydantic-settings==2.1.0
annotated-types==0.7.0
typing_extensions==4.15.0

# ===============================
# Database & ORM
# ===============================
SQLAlchemy==2.0.46
alembic==1.18.1
pyodbc==5.0.1
greenlet==3.3.0

# ===============================
# Vector Databases
# ===============================
faiss-cpu==1.8.0
qdrant-client==1.7.0
portalocker==2.10.1

# ===============================
# AI / ML / NLP
# ===============================
openai==2.15.0
sentence-transformers==5.2.0
transformers==4.57.6
huggingface-hub==0.36.0
torch==2.10.0
tokenizers==0.22.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
numpy==1.26.4
regex==2026.1.15
joblib==1.5.3
threadpoolctl==3.6.0

# ===============================
# Data Processing & Visualization
# ===============================
pandas==2.3.3
matplotlib==3.10.8
pillow==12.1.0
sympy==1.14.0
networkx==3.6.1

# ===============================
# HTTP & Networking
# ===============================
requests==2.31.0
httpx==0.28.1
httpcore==1.0.9
urllib3==1.26.20
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
sniffio==1.3.1
anyio==4.12.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
grpcio==1.76.0
grpcio-tools==1.76.0

# ===============================
# Scheduling & Background Jobs
# ===============================
APScheduler==3.11.2
tzlocal==5.3.1
pytz==2025.2
tzdata==2025.3

# ===============================
# Utilities & Logging
# ===============================
python-dotenv==1.0.0
python-json-logger==2.0.7
tqdm==4.67.1
filelock==3.20.3
platformdirs==4.5.1
packaging==25.0
psutil==7.2.1
PyYAML==6.0.3
fsspec==2026.1.0
distro==1.9.0
six==1.17.0

# ===============================
# Testing
# ===============================
pytest==7.4.3
pytest-flask==1.3.0
iniconfig==2.3.0
pluggy==1.6.0

# ===============================
# Dev / Interactive (optional)
# ===============================
ipython==9.9.0
ipykernel==7.1.0
jupyter-client==8.8.0
jupyter-core==5.9.1
traitlets==5.14.3
prompt-toolkit==3.0.52
Pygments==2.19.2
colorama==0.4.6
wcwidth==0.2.14
